feat: rename rockappreact to brownfieldlib
