+++ apps/AndroidApp/app/src/rock/java/com/callstack/brownfield/android/example/BrownfieldStore.kt
@@ -1,0 +1,4 @@
+package com.callstack.brownfield.android.example
+
+typealias BrownfieldStore = com.callstack.rnbrownfield.demo.rockapp.BrownfieldStore
+typealias User = com.callstack.rnbrownfield.demo.rockapp.User
