+++ apps/AndroidApp/app/src/rock/java/com/callstack/brownfield/android/example/ReactNativeConstants.kt
@@ -1,0 +1,6 @@
+package com.callstack.brownfield.android.example
+
+object ReactNativeConstants {
+    const val MAIN_MODULE_NAME = "RockApp"
+    const val APP_NAME = "Android"
+}
