+++ apps/RockApp/android/app/build.gradle
@@ -77,9 +77,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.rockapp"
+    namespace "com.callstack.rnbrownfield.demo.rockapp"
     defaultConfig {
-        applicationId "com.rockapp"
+        applicationId "com.callstack.rnbrownfield.demo.rockapp"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
