+++ apps/RockApp/android/app/src/main/java/com/callstack/rnbrownfield/demo/rockapp/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.rockapp
+package com.callstack.rnbrownfield.demo.rockapp
 
 import android.app.Application
 import com.facebook.react.PackageList
