+++ apps/RockApp/android/settings.gradle
@@ -4,4 +4,4 @@ extensions.configure(com.facebook.react.ReactSettingsExtension){ ex -> ex.autoli
 rootProject.name = 'RockApp'
 include ':app'
 includeBuild('../node_modules/@react-native/gradle-plugin')
-include(":rockappreact")
+include(":brownfieldlib")
