+++ apps/RockApp/ios/BrownfieldLib/BrownfieldLib.h
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+
+//! Project version number for BrownfieldLib.
+FOUNDATION_EXPORT double BrownfieldLibVersionNumber;
+
+//! Project version string for BrownfieldLib.
+FOUNDATION_EXPORT const unsigned char BrownfieldLibVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <BrownfieldLib/PublicHeader.h>
