+++ apps/RockApp/ios/Podfile
@@ -25,7 +25,7 @@ target 'RockApp' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'RockAppReact' do
+  target 'BrownfieldLib' do
     inherit! :complete
   end
