--- apps/RockApp/ios/RockAppReact/RockAppReact.h
@@ -1,9 +1,0 @@
-#import <Foundation/Foundation.h>
-
-//! Project version number for RockAppReact.
-FOUNDATION_EXPORT double RockAppReactVersionNumber;
-
-//! Project version string for RockAppReact.
-FOUNDATION_EXPORT const unsigned char RockAppReactVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <RockAppReact/PublicHeader.h>
